+++ FileZilla Server/trunk/FileZilla Server/source/Interface/OptionsGeneralPage.cpp
@@ -144,12 +144,12 @@ BOOL COptionsGeneralPage::IsDataValid()
 
 void COptionsGeneralPage::LoadData()
 {
-	m_Port = m_pOptionsDlg->GetOption(OPTION_SERVERPORT);	
-	m_Threadnum.Format("%d", m_pOptionsDlg->GetOptionVal(OPTION_THREADNUM));	
-	m_MaxUsers.Format("%d", m_pOptionsDlg->GetOptionVal(OPTION_MAXUSERS));	
-	m_Timeout.Format("%d", m_pOptionsDlg->GetOptionVal(OPTION_TIMEOUT));	
-	m_NoTransferTimeout.Format("%d", m_pOptionsDlg->GetOptionVal(OPTION_NOTRANSFERTIMEOUT));
-	m_LoginTimeout.Format("%d", m_pOptionsDlg->GetOptionVal(OPTION_LOGINTIMEOUT));
+	m_Port = m_pOptionsDlg->GetOption(OPTION_SERVERPORT);
+	m_Threadnum.Format("%d", static_cast<int>(m_pOptionsDlg->GetOptionVal(OPTION_THREADNUM)));
+	m_MaxUsers.Format("%d", static_cast<int>(m_pOptionsDlg->GetOptionVal(OPTION_MAXUSERS)));
+	m_Timeout.Format("%d", static_cast<int>(m_pOptionsDlg->GetOptionVal(OPTION_TIMEOUT)));
+	m_NoTransferTimeout.Format("%d", static_cast<int>(m_pOptionsDlg->GetOptionVal(OPTION_NOTRANSFERTIMEOUT)));
+	m_LoginTimeout.Format("%d", static_cast<int>(m_pOptionsDlg->GetOptionVal(OPTION_LOGINTIMEOUT)));
 }
 
 void COptionsGeneralPage::SaveData()
